### Related Issue
https://issues.redhat.com/browse/ACM-5899

### Description of changes
- **Reduced `DB_MAX_CONNS` from 20 to 10.** 
    - Should reduce memory usage because idle connections consume memory on postgres and indexer pods.
    - I haven't noticed a performance impact from reducing these connections.
- **Increased `DB_BATCH_SIZE` from 500 to 2500.** 
    - With a larger batch size, less connections are required. With this change only 2 connections are used to process a request to add 5,000 resources. Without this it would use 10 connections.
- **Increased `DB_MIN_CONNS` from 0 to 2.**
    - Reduces the time needed to process the first request.
- **Reduced `REQUEST_LIMIT` from 50 to 25.**
    - Reduces the max memory needed to process all concurrent requests.
    - Increases the chances that an individual requests will complete without timing out.

### Test results
**Scenario:** Used Locust to simulate 50 clusters with 100k resources every 1 second.

Measured | Before change | After change
---             | ---                    | ---
Max indexer memory (on local dev env) | 12.49 GB | 7.58 GB
Max request time        | 248s | 120s
Average request time | 154s | 41s

The total time to process all 50 clusters wasn't measured, but observed to be similar in both scenarios, about 5 minutes.