@SherinV 2023-09-18
We should check how it performs in the next scale run.
/lgtm
/hold