@jlpadilla 2023-09-18
Added test metrics to the description.

### Test results
**Scenario:** Used Locust to simulate 50 clusters with 100k resources every 1 second.

Measured | Before change | After change
---             | ---                    | ---
Max indexer memory (on local dev env) | 12.49 GB | 7.58 GB
Max request time        | 248s | 120s
Average request time | 154s | 41s